+++ website/components/Layout.tsx
@@ -1,4 +1,4 @@
-import { Link, useMatchRoute, useMatches } from "@tanstack/react-router"
+import { Link, useMatchRoute } from "@tanstack/react-router"
 import { allArticles } from "content-collections"
 import { BookOpen, ExternalLink, FileText } from "lucide-react"
 import { articleIcons } from "../helpers/article-icon.ts"
@@ -114,23 +114,5 @@ function AppSidebar() {
       </SidebarFooter>
       <SidebarRail />
     </Sidebar>
-  )
-}
-
-function Breadcrumbs() {
-  const matches = useMatches()
-  const last = matches[matches.length - 1]
-  const loaderData = last?.loaderData as { title?: string } | undefined
-
-  return (
-    <nav className="flex items-center gap-1.5 text-sm text-muted-foreground">
-      <Link to="/" className="hover:text-foreground transition-colors">
-        Docs
-      </Link>
-      <span>/</span>
-      {loaderData?.title && (
-        <span className="text-foreground font-medium">{loaderData.title}</span>
-      )}
-    </nav>
   )
 }
